皇家坐骑修正.背包栏扩（大）充修改

+++ gms-server/src/main/java/org/gms/client/inventory/Inventory.java
@@ -53,10 +53,12 @@ public class Inventory implements Iterable<Item> {
     protected final Lock lock = new ReentrantLock(true);
 
     protected Character owner;
-    protected byte slotLimit;
+  //  protected byte slotLimit;
+    protected short slotLimit;
     protected boolean checked = false;
 
-    public Inventory(Character mc, InventoryType type, byte slotLimit) {
+    public Inventory(Character mc, InventoryType type, short slotLimit) {
+  //  public Inventory(Character mc, InventoryType type, byte slotLimit) {
         this.owner = mc;
         this.inventory = new LinkedHashMap<>();
         this.type = type;
@@ -71,7 +73,8 @@ public class Inventory implements Iterable<Item> {
         return type.equals(InventoryType.EQUIP) || type.equals(InventoryType.EQUIPPED);
     }
 
-    public byte getSlotLimit() {
+    public short getSlotLimit() {
+       // public byte getSlotLimit()
         lock.lock();
         try {
             return slotLimit;
@@ -96,7 +99,8 @@ public class Inventory implements Iterable<Item> {
                 }
             }
 
-            slotLimit = (byte) newLimit;
+          //  slotLimit = (byte) newLimit;
+            slotLimit = (short) newLimit;
         } finally {
             lock.unlock();
         }
